util refactor
