+++ src/frontend/src/components/utils/browserLogos.ts
